+++ src/test/java/com/siemens/pki/cmpracomponent/test/TestCentralKeyGenerationWithKeyAgreement.java
@@ -340,7 +340,8 @@ public class TestCentralKeyGenerationWithKeyAgreement extends CkgOnlineEnrollmen
                             final byte[] transactionID,
                             final String requesterDn,
                             final byte[] certTemplate,
-                            final String requestedSubjectDn) {
+                            final String requestedSubjectDn,
+                            byte[] pkiMessage) {
                         LOGGER.debug(
                                 "checkAndModifyCertRequest called with transactionID: {}, requesterDn: {}, requestedSubjectDn: {}",
                                 new BigInteger(transactionID),
@@ -365,7 +366,8 @@ public class TestCentralKeyGenerationWithKeyAgreement extends CkgOnlineEnrollmen
                             final byte[] transactionID,
                             final String requesterDn,
                             final byte[] pkcs10CertRequest,
-                            final String requestedSubjectDn) {
+                            final String requestedSubjectDn,
+                            byte[] pkiMessage) {
                         LOGGER.debug(
                                 "checkP10CertRequest called with transactionID: {}, requesterDn: {}, requestedSubjectDn: {}",
                                 new BigInteger(transactionID),