+++ src/test/java/com/siemens/pki/cmpracomponent/test/TestCentralKeyGenerationWithKeyTransport.java
@@ -253,7 +253,8 @@ public class TestCentralKeyGenerationWithKeyTransport extends CkgOnlineEnrollmen
                             final byte[] transactionID,
                             final String requesterDn,
                             final byte[] certTemplate,
-                            final String requestedSubjectDn) {
+                            final String requestedSubjectDn,
+                            byte[] pkiMessage) {
                         LOGGER.debug(
                                 "checkAndModifyCertRequest called with transactionID: {}, requesterDn: {}, requestedSubjectDn: {}",
                                 new BigInteger(transactionID),
@@ -278,7 +279,8 @@ public class TestCentralKeyGenerationWithKeyTransport extends CkgOnlineEnrollmen
                             final byte[] transactionID,
                             final String requesterDn,
                             final byte[] pkcs10CertRequest,
-                            final String requestedSubjectDn) {
+                            final String requestedSubjectDn,
+                            byte[] pkiMessage) {
                         LOGGER.debug(
                                 "checkP10CertRequest called with transactionID: {}, requesterDn: {}, requestedSubjectDn: {}",
                                 new BigInteger(transactionID),